% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_plots.R
\name{make_gsea_bar}
\alias{make_gsea_bar}
\title{Make bar plot with gene set enrichment results}
\usage{
make_gsea_bar(
  data,
  enrich_var = NULL,
  p_var = "p_value",
  dir = "both",
  color_by = "pval",
  x_by = "enrich",
  n_shown = 10,
  sig_only = FALSE,
  term_var = "term",
  remove_collection_name = FALSE,
  term_words_per_line = 2,
  term_max_lines = 4,
  theme = NULL
)
}
\arguments{
\item{data}{a table containing gene set enrichment results.}

\item{enrich_var}{name of the column containing enrichment values. Defaults to "NES" or "odds_ratio".}

\item{p_var}{name of the column containing p-values. Defaults to "p_value".}

\item{dir}{direction(s) to plot. Options are ("both","pos","neg").}

\item{color_by}{variable to color points by. Options are ("pval","dir","enrich").}

\item{x_by}{variable to plot on x-axis. Options are ("pval","enrich").}

\item{n_shown}{number of terms to show.}

\item{sig_only}{whether to only show significantly enriched terms.}

\item{term_var}{name of the column containing the term. Defaults to "term".}

\item{remove_collection_name}{whether to remover the gene set collection name e.g. HALLMARK.}

\item{term_words_per_line}{number of words per line in term.}

\item{term_max_lines}{maximum number of lines the term can use.}

\item{theme}{ggplot theme applied to the plot.}
}
\description{
Make bar plot with gene set enrichment results
}
