% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_themes.R
\name{scale_color_Publication}
\alias{scale_color_Publication}
\title{Set publication color scheme}
\usage{
scale_color_Publication(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Set publication color scheme
}
