% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_themes.R
\name{scale_fill_Publication}
\alias{scale_fill_Publication}
\title{Set publication fill scheme}
\usage{
scale_fill_Publication(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Set publication fill scheme
}
