% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_themes.R
\name{theme_publication}
\alias{theme_publication}
\title{Set publication theme on ggplot object}
\usage{
theme_publication(
  base_size = 12,
  base_family = "Helvetica",
  base_face = "bold",
  legend_bottom = TRUE
)
}
\arguments{
\item{legend_bottom}{}
}
\description{
Set publication theme on ggplot object
}
