% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_themes.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Set publication theme on ggplot object}
\usage{
theme_Publication(
  base_size = 12,
  base_family = "Helvetica",
  base_face = "bold",
  legend_bottom = TRUE
)
}
\arguments{
\item{base_family}{}
}
\value{

}
\description{
Set publication theme on ggplot object
}
\examples{
https://rpubs.com/Koundy/71792
}
